# monolito/openapi.yaml
openapi: 3.0.3
info:
  title: E-commerce Monolito API para ADI
  description: API REST para backend monolítico de mini e-commerce
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Servidor local

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: demo@demo.com
        password:
          type: string
          example: demo1234

    LoginResponse:
      type: object
      properties:
        token:
          type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        user:
          type: object
          properties:
            id:
              type: integer
              example: 1
            email:
              type: string
              example: demo@demo.com

    MeResponse:
      type: object
      properties:
        user:
          type: object
          properties:
            id:
              type: integer
              example: 1
            email:
              type: string
              example: demo@demo.com

    Product:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: Laptop Lenovo ThinkPad
        priceCents:
          type: integer
          description: Precio en centavos
          example: 89999
        stock:
          type: integer
          example: 15

    ProductListResponse:
      type: object
      properties:
        products:
          type: array
          items:
            $ref: '#/components/schemas/Product'

    ProductResponse:
      type: object
      properties:
        product:
          $ref: '#/components/schemas/Product'

    OrderItem:
      type: object
      properties:
        productId:
          type: integer
          example: 1
        unitPriceCents:
          type: integer
          example: 89999
        quantity:
          type: integer
          example: 2

    Order:
      type: object
      properties:
        id:
          type: integer
          example: 1
        totalCents:
          type: integer
          description: Total del pedido en centavos
          example: 179998
        createdAt:
          type: string
          format: date-time
          example: "2026-01-10T18:54:02.000Z"
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'

    OrderListResponse:
      type: object
      properties:
        orders:
          type: array
          items:
            $ref: '#/components/schemas/Order'

    CreateOrderRequest:
      type: object
      required:
        - items
      properties:
        items:
          type: array
          items:
            type: object
            required:
              - productId
              - quantity
            properties:
              productId:
                type: integer
                example: 1
              quantity:
                type: integer
                example: 2

    CreateOrderResponse:
      type: object
      properties:
        order:
          $ref: '#/components/schemas/Order'

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: Error interno del servidor

    HealthResponse:
      type: object
      properties:
        ok:
          type: boolean
          example: true
        service:
          type: string
          example: monolito

paths:
  /health:
    get:
      summary: Health check
      tags:
        - Health
      responses:
        '200':
          description: Servicio funcionando correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'

  /login:
    post:
      summary: Autenticación de usuario
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login exitoso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          description: Input inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Credenciales inválidas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Error interno
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /me:
    get:
      summary: Obtener usuario autenticado
      tags:
        - Auth
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Usuario obtenido correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeResponse'
        '401':
          description: Token ausente o inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Error interno
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /products:
    get:
      summary: Listar todos los productos
      tags:
        - Catálogo
      responses:
        '200':
          description: Lista de productos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductListResponse'
        '500':
          description: Error interno
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /products/{id}:
    get:
      summary: Obtener un producto por ID
      tags:
        - Catálogo
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: ID del producto
      responses:
        '200':
          description: Producto encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '400':
          description: Producto no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Error interno
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /orders:
    get:
      summary: Listar pedidos del usuario autenticado
      tags:
        - Pedidos
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de pedidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderListResponse'
        '401':
          description: Token ausente o inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Error interno
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      summary: Crear un nuevo pedido
      tags:
        - Pedidos
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
      responses:
        '201':
          description: Pedido creado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateOrderResponse'
        '400':
          description: Input inválido o producto inexistente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Token ausente o inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Error interno
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'